requirements-docker.txt
# User requested speed over functionality for visual check
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
yfinance>=0.2.18
feedparser>=6.0.10
pydantic>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
aiosqlite>=0.19.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
loguru>=0.7.0
APScheduler>=3.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
ta>=0.10.2
textblob>=0.17.1
scipy>=1.11.0
websockets>=11.0.0
# ollama>=0.1.7
# sentence-transformers>=2.2.2
